Extract generator from gui
Avoid tkinter import where it's not needed
